requirements.dev.txt
-r requirements.txt
flake8==3.9.2
django-debug-toolbar==3.2.2
django-extensions==3.1.3
pyparsing==2.4.7
pydot==1.4.2
django-webtest==1.9.7
factory-boy==2.8.1
django-celery-results